moved common variables and updateCell to a separate file
