+++ src/scripts/handleFormSubmission.js
@@ -1,8 +1,8 @@
 import { formatCurrency } from './formatCurrency';
 import { pushNotification } from './pushNotification';
+import { tbody } from './variables';
 
 const addDataToTable = (employeInfo) => {
-  const tbody = document.querySelector('tbody');
   const tr = document.createElement('tr');
 
   Object.values(employeInfo).forEach((cellData) => {
